Make name semi-persistent and modify collisions
The name persists now between the home and the gameplay pages, however is has yet to persist between the home and the other pages (skins, leaderboard, settings, upgrades). The collisions are modified so now the snake can exist on the edges, however it brings back the bug of it going one step too far into the bounds.